refactor pb-manage CLI into modular structure
