+++ pb-manage/bin/pb-manage.js
@@ -1,284 +1,10 @@
 #!/usr/bin/env node
-const fs = require('fs');
-const path = require('path');
-const { execSync } = require('child_process');
-const minimist = require('minimist');
-
-const argv = minimist(process.argv.slice(2), {
-  string: ['config', 'ssh-key', 'env'],
-  boolean: ['local', 'remote'],
-  alias: { c: 'config', k: 'ssh-key', e: 'env' }
-});
-
-function loadConfig() {
-  const cfgPath = argv.config
-    ? path.resolve(argv.config)
-    : path.join(process.cwd(), 'pb.config.json');
-  if (!fs.existsSync(cfgPath)) {
-    console.error('Missing pb.config.json, run `pb-manage init`');
-    process.exit(1);
-  }
-  const cfg = JSON.parse(fs.readFileSync(cfgPath, 'utf8'));
-  if (argv['ssh-key']) cfg.sshKey = argv['ssh-key'];
-  return cfg;
-}
-
-function run(cmd) {
-  execSync(cmd, { stdio: 'inherit', shell: true });
-}
-
-function ssh(cmd) {
-  const cfg = loadConfig();
-  const keyPath = argv['ssh-key'] || cfg.sshKey;
-  const key = keyPath ? `-i ${keyPath}` : '';
-  run(`ssh ${key} ${cfg.sshUser}@${cfg.vpsHost} '${cmd}'`);
-}
-
-function scpFromRemote(remotePath, localPath) {
-  const cfg = loadConfig();
-  const keyPath = argv['ssh-key'] || cfg.sshKey;
-  const key = keyPath ? `-i ${keyPath}` : '';
-  run(`scp ${key} ${cfg.sshUser}@${cfg.vpsHost}:${remotePath} ${localPath}`);
-}
-
-function scpToRemote(localPath, remotePath) {
-  const cfg = loadConfig();
-  const keyPath = argv['ssh-key'] || cfg.sshKey;
-  const key = keyPath ? `-i ${keyPath}` : '';
-  run(`scp ${key} ${localPath} ${cfg.sshUser}@${cfg.vpsHost}:${remotePath}`);
-}
+const commands = require('../lib/commands');
+const { argv, init, dev, migrate, seed, create, destroy, backup, restore, pull, deploy } = commands;
 
 function usage() {
   console.log('Usage: pb-manage [--config path] [--ssh-key key] [--env name] <command> [options]');
-  console.log('Commands: init, dev, migrate, seed, create, destroy, backup, restore, deploy');
-}
-
-function ensureDir(p) {
-  if (!fs.existsSync(p)) fs.mkdirSync(p, { recursive: true });
-}
-
-function writeFileIfMissing(file, content) {
-  if (!fs.existsSync(file)) {
-    fs.writeFileSync(file, content);
-    console.log('created', file);
-  }
-}
-
-function init() {
-  ensureDir('pb_migrations');
-  ensureDir('pb_seeds');
-  writeFileIfMissing('pb.config.json', JSON.stringify({
-    vpsHost: 'your.vps.host',
-    sshUser: 'root',
-    domain: 'api.example.com'
-  }, null, 2));
-
-  writeFileIfMissing('Dockerfile', `FROM alpine:latest
-ARG PB_VERSION=0.28.3
-RUN apk add --no-cache unzip ca-certificates
-ADD https://github.com/pocketbase/pocketbase/releases/download/v\${PB_VERSION}/pocketbase_\${PB_VERSION}_linux_amd64.zip /tmp/pb.zip
-RUN unzip /tmp/pb.zip -d /pb/
-WORKDIR /pb
-COPY pb_migrations /pb/pb_migrations
-COPY pb_seeds /pb/pb_seeds
-EXPOSE 8090
-CMD ["./pocketbase", "serve", "--http=0.0.0.0:8090"]
-`);
-
-  writeFileIfMissing('docker-compose.yml', `version: '3.9'
-services:
-  pocketbase:
-    build: .
-    ports:
-      - "8090:8090"
-    volumes:
-      - ./pb_data:/pb/pb_data
-      - ./pb_migrations:/pb/pb_migrations
-      - ./pb_seeds:/pb/pb_seeds
-    environment:
-      - PB_ENV=development
-  nginx:
-    image: nginx:alpine
-    ports:
-      - "80:80"
-    volumes:
-      - ./nginx.conf:/etc/nginx/conf.d/default.conf:ro
-`);
-
-  writeFileIfMissing('nginx.conf', `server {
-    listen 80;
-    server_name localhost;
-    location / {
-        proxy_pass http://pocketbase:8090;
-        proxy_set_header Host $host;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    }
-}
-`);
-
-  ensureDir('.github/workflows');
-  writeFileIfMissing('.github/workflows/pb-deploy.yml', `name: Deploy PocketBase
-
-on:
-  push:
-    branches: [main]
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    env:
-      REGISTRY: ghcr.io
-      IMAGE: \${{ env.REGISTRY }}/\${{ github.repository }}:latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version: '18'
-      - uses: docker/setup-buildx-action@v3
-      - uses: docker/login-action@v3
-        with:
-          registry: \${{ env.REGISTRY }}
-          username: \${{ secrets.REGISTRY_USERNAME }}
-          password: \${{ secrets.REGISTRY_PASSWORD }}
-      - uses: docker/build-push-action@v5
-        with:
-          context: .
-          push: true
-          tags: \${{ env.IMAGE }}
-      - name: Deploy via pb-manage
-        env:
-          SSH_KEY: \${{ secrets.SSH_KEY }}
-        run: |
-          echo "$SSH_KEY" > key.pem
-          chmod 600 key.pem
-          node pb-manage/bin/pb-manage.js deploy production --ssh-key key.pem
-          rm key.pem
-`);
-
-  writeFileIfMissing('README.md', `# PocketBase Project
-
-Run \`pb-manage dev\` to start development server.
-`);
-}
-
-function ensureDockerRunning() {
-  try {
-    execSync('docker info', { stdio: 'ignore' });
-  } catch (err) {
-    console.error('Docker does not appear to be running. Please start Docker.');
-    process.exit(1);
-  }
-}
-
-function dev() {
-  ensureDockerRunning();
-  run('docker-compose up -d --build');
-  migrate();
-  seed();
-  run('docker-compose logs -f');
-}
-
-function migrate() {
-  run('docker-compose exec pocketbase pocketbase migrate up');
-}
-
-function seed() {
-  run('docker-compose exec pocketbase pocketbase seed up');
-}
-
-function create(env = argv.env) {
-  if (!env) return console.error('Missing environment name');
-  const cfg = loadConfig();
-  // Clone or restore the base data directory if it exists
-  ssh(`if [ -d ~/pb_base/pb_data ]; then rm -rf ~/pb_data/${env} && cp -r ~/pb_base/pb_data ~/pb_data/${env}; else mkdir -p ~/pb_data/${env}; fi`);
-  // Ensure a Docker network for nginx routing
-  ssh('docker network inspect pb-net >/dev/null 2>&1 || docker network create pb-net');
-  // Start the PocketBase container on the network
-  ssh(`docker run -d --name pb-${env} --network pb-net -v ~/pb_data/${env}:/pb/pb_data pocketbase`);
-  // Apply migrations and seeds on the new container
-  ssh(`docker exec pb-${env} pocketbase migrate up`);
-  ssh(`docker exec pb-${env} pocketbase seed up`);
-  // Configure nginx to expose the instance
-  const domain = cfg.domain;
-  ssh(`cat <<'EOF' > /etc/nginx/conf.d/pb-${env}.conf
-server {
-    listen 80;
-    server_name ${env}.${domain};
-    location / {
-        proxy_pass http://pb-${env}:8090;
-        proxy_set_header Host $host;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    }
-}
-EOF`);
-  ssh('nginx -s reload');
-}
-
-function destroy(env = argv.env) {
-  if (!env) return console.error('Missing environment name');
-  ssh(`docker rm -f pb-${env}`);
-  ssh(`rm -rf ~/pb_data/${env}`);
-  ssh(`rm -f /etc/nginx/conf.d/pb-${env}.conf`);
-  ssh('nginx -s reload');
-}
-
-function backup(env = argv.env) {
-  if (!env) return console.error('Missing environment name');
-  const remoteZip = `/tmp/${env}-backup.zip`;
-  ssh(`zip -r ${remoteZip} ~/pb_data/${env}`);
-  if (argv.local) {
-    scpFromRemote(remoteZip, `./${env}-backup.zip`);
-    if (!argv.remote) ssh(`rm -f ${remoteZip}`);
-  } else {
-    console.log(`Backup stored on server at ${remoteZip}`);
-  }
-}
-
-function restore(env = argv.env, file) {
-  if (!env || !file) return console.error('Usage: pb-manage restore <env> <file>');
-  if (!fs.existsSync(file)) return console.error(`File not found: ${file}`);
-
-  const remoteZip = `/tmp/${path.basename(file)}`;
-  scpToRemote(file, remoteZip);
-
-  ssh(`docker stop pb-${env} || true`);
-  ssh(`unzip -o ${remoteZip} -d ~/pb_data/${env}`);
-  ssh(`docker start pb-${env} || docker run -d --name pb-${env} --network pb-net -v ~/pb_data/${env}:/pb/pb_data pocketbase`);
-  ssh(`rm -f ${remoteZip}`);
-}
-
-function pull(env = argv.env) {
-  if (!env) return console.error('Missing environment name');
-  ensureDockerRunning();
-
-  const remoteZip = `/tmp/${env}-backup.zip`;
-  const localZip = `./${env}-backup.zip`;
-
-  // Create backup on the remote host and download it
-  ssh(`zip -r ${remoteZip} ~/pb_data/${env}`);
-  scpFromRemote(remoteZip, localZip);
-  ssh(`rm -f ${remoteZip}`);
-
-  // Stop local container if running
-  try { run('docker-compose down'); } catch (err) {}
-
-  // Restore the backup into ./pb_data
-  run('rm -rf pb_data');
-  run(`unzip -o ${localZip}`);
-  run(`[ -d root/pb_data/${env} ] && mv root/pb_data/${env} pb_data && rm -rf root`);
-  run(`[ -d pb_data/${env} ] && mv pb_data/${env} pb_data`);
-
-  // Restart the local environment
-  run('docker-compose up -d --build');
-}
-
-function deploy(env = argv.env) {
-  if (!env) return console.error('Missing environment name');
-  create(env);
-  migrate();
-  seed();
+  console.log('Commands: init, dev, migrate, seed, create, destroy, backup, restore, pull, deploy');
 }
 
 if (require.main === module) {
@@ -296,26 +22,5 @@ if (require.main === module) {
   else if (cmd === 'deploy') deploy(argv._[1]);
   else usage();
 } else {
-  module.exports = {
-    argv,
-    loadConfig,
-    run,
-    ssh,
-    scpFromRemote,
-    scpToRemote,
-    ensureDir,
-    writeFileIfMissing,
-    ensureDockerRunning,
-    init,
-    dev,
-    migrate,
-    seed,
-    create,
-    destroy,
-    backup,
-    restore,
-    pull,
-    deploy,
-    usage,
-  };
+  module.exports = { ...commands, usage };
 }
