+++ pb-manage/lib/commands.js
@@ -1,0 +1,27 @@
+const argv = require('./argv');
+const { loadConfig, run, ssh, scpFromRemote, scpToRemote, ensureDir, writeFileIfMissing, ensureDockerRunning } = require('./utils');
+const { init } = require('./init');
+const { migrate, seed, dev, pull } = require('./local');
+const { create, destroy, backup, restore, deploy } = require('./remote');
+
+module.exports = {
+  argv,
+  loadConfig,
+  run,
+  ssh,
+  scpFromRemote,
+  scpToRemote,
+  ensureDir,
+  writeFileIfMissing,
+  ensureDockerRunning,
+  init,
+  migrate,
+  seed,
+  dev,
+  pull,
+  create,
+  destroy,
+  backup,
+  restore,
+  deploy,
+};
